requirements.txt
fastapi==0.116.1
pandas==2.3.1
pydantic==2.11.7
pytz==2024.1
Requests==2.32.4
SQLAlchemy==2.0.36
uvicorn==0.35.0
pyodbc==4.0.39
python-dotenv==1.0.1